検証用カードの追加とマナカーブ調整NPCの検証
・検証用のカードを追加
・マナカーブを調整するNPCの検証中

+++ Assets/Script/Card/CardEntity.cs
@@ -11,6 +11,6 @@ public class CardEntity : ScriptableObject
     public int cost;
     public int power;
     public int hp;
-    public int evaluation;
+    public float evaluation;
     //public Sprite icon; //‰æ‘œ•\¦iŒã“ú’Ç‰Á?j
 }
